requirements.txt
Django==1.11.7
django-braces==1.11.0
django-crispy-forms==1.7.0
django-model-utils==3.0.0
django-turbolinks==0.5.1
olefile==0.44
packaging==16.8
Pillow==4.3.0
pyparsing==2.2.0
python-decouple==3.1
pytz==2017.3

